+++ NEWS.md
@@ -1,5 +1,13 @@
 # Package updates
 
+## tbcleanr 0.1.2.0
+### 20 Jan 2019
+
+* Added txhistory() to clean variables for previous treatment history
+* 'cdhistory' variable included in`adm_subset.koch6()`
+
+---
+
 ## tbcleanr 0.1.1.7
 ### 8 Jan 2019